requirements.txt
accelerate==0.25.0
pandas==2.1.4
numpy==1.26.2
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
diffusers==0.24.0
pillow